feat: strf-10477, strf-10457 Base-level rules and conditional multi file import
