+++ .github/workflows/pull_request_review.yml
@@ -5,7 +5,7 @@ name: Tests
 
 on:
     pull_request:
-        branches: [master, main]
+        branches: [master, main, release-**]
 
 jobs:
     build:
