+++ lib/bundle-validator.js
@@ -47,7 +47,7 @@ class BundleValidator {
         this.validationTasks = [
             this._validateThemeConfiguration.bind(this),
             this._validateThemeSchema.bind(this),
-            // this._validateSchemaTranslations.bind(this),
+            this._validateSchemaTranslations.bind(this),
             this._validateTemplatesFrontmatter.bind(this),
             this._validateCssFiles.bind(this),
         ];
